+++ libs/engine/Cargo.toml
@@ -38,6 +38,7 @@ jwt_auth_plugin = { path = "../../plugins/jwt_auth" }
 federation_query_planner = { path = "../../libs/federation_query_planner" }
 opentelemetry_plugin = { path = "../../plugins/opentelemetry" }
 minitrace = { workspace = true }
+minitrace_reqwest = { path = "../minitrace_reqwest" }
 
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]
 tokio = { workspace = true }
