+++ libs/federation_query_planner/Cargo.toml
@@ -19,6 +19,7 @@ linked-hash-map = "0.5.6"
 futures = { workspace = true }
 async-graphql = { version = "6.0.11", features = ["dynamic-schema"] }
 minitrace = { workspace = true }
+minitrace_reqwest = { path = "../minitrace_reqwest" }
 
 [dev-dependencies]
 insta = { version = "1.29.0", features = ["yaml", "json"] }
