+++ libs/minitrace_reqwest/Cargo.toml
@@ -9,6 +9,7 @@ path = "src/lib.rs"
 [dependencies]
 minitrace = { workspace = true }
 reqwest-middleware = { workspace = true }
+conductor_tracing = { path = "../tracing" }
 reqwest = { workspace = true }
 async-trait = { workspace = true }
 task-local-extensions = "0.1.4"
