+++ libs/tracing/src/lib.rs
@@ -1,10 +1,7 @@
 pub mod config;
 pub mod manager;
 pub mod minitrace_mgr;
+pub mod otel_attrs;
 pub mod otel_utils;
-pub mod reqwest_utils;
 #[cfg(target_arch = "wasm32")]
 pub mod wasm_span_processor;
-
-pub use opentelemetry::trace::FutureExt;
-pub use tracing_opentelemetry::OpenTelemetrySpanExt;
